requirements.txt
pytest
flask==3.0.0
numpy
scikit-learn==1.3.1
joblib==1.3.2
Pillow==10.0.1
